+++ src/Captura.Base/Alignment.cs
