+++ src/Captura.Base/MouseOverlaySettings.cs
