+++ src/Captura.Base/PositionedOverlaySettings.cs
