+++ src/Captura.Base/PropertyStore.cs
