+++ src/Captura.Base/TextOverlaySettings.cs
