+++ src/Captura.Console/Captura.Console.csproj
@@ -71,6 +71,10 @@
       <Project>{f8568b22-8ac6-49c5-92df-e1c359ff1fab}</Project>
       <Name>Captura.Core</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Captura.MouseKeyHook\Captura.MouseKeyHook.csproj">
+      <Project>{6CC6E9C8-9A94-4FB0-8396-47E2ECBD8300}</Project>
+      <Name>Captura.MouseKeyHook</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Captura.SharpAvi\Captura.SharpAvi.csproj">
       <Project>{bb8ff9f8-ec60-4a4a-a3f0-d9fbfb585d86}</Project>
       <Name>Captura.SharpAvi</Name>
