+++ src/Captura.Core/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="MouseKeyHook" version="5.6.0" targetFramework="net461" />
   <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net461" />
   <package id="Ookii.Dialogs.WindowsForms" version="1.0" targetFramework="net461" />
 </packages>
