+++ src/Captura.MouseKeyHook/DummyKeyRecord.cs
