+++ src/Captura.MouseKeyHook/IKeyRecord.cs
