+++ src/Captura.MouseKeyHook/KeyRecord.cs
