+++ src/Captura.MouseKeyHook/KeyRecords.cs
