+++ src/Captura.MouseKeyHook/KeystrokesSettings.cs
