+++ src/Captura.MouseKeyHook/MouseClickSettings.cs
