+++ src/Captura.MouseKeyHook/RepeatKeyRecord.cs
