+++ src/Captura.MouseKeyHook/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MouseKeyHook" version="5.6.0" targetFramework="net461" />
+</packages>
