+++ src/Captura/Captura.csproj
@@ -426,6 +426,10 @@
       <Project>{f8568b22-8ac6-49c5-92df-e1c359ff1fab}</Project>
       <Name>Captura.Core</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Captura.MouseKeyHook\Captura.MouseKeyHook.csproj">
+      <Project>{6CC6E9C8-9A94-4FB0-8396-47E2ECBD8300}</Project>
+      <Name>Captura.MouseKeyHook</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Captura.Native\Captura.Native.csproj">
       <Project>{81f2e68e-4708-4d52-8419-51888f22d05e}</Project>
       <Name>Captura.Native</Name>
