+++ src/Captura/Controls/TextOverlaySettingsControl.xaml
@@ -6,8 +6,9 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:captura="clr-namespace:Captura;assembly=Captura.Core"
+             xmlns:base="clr-namespace:Captura;assembly=Captura.Base"
              mc:Ignorable="d"
-             d:DataContext="{d:DesignInstance captura:TextOverlaySettings}">
+             d:DataContext="{d:DesignInstance base:TextOverlaySettings}">
     <StackPanel>
         <Grid>
             <Grid.ColumnDefinitions>
